+++ camera/QCamera2/HAL3/QCamera3Channel.cpp
@@ -527,7 +527,7 @@ void QCamera3Channel::dumpYUV(mm_camera_buf_def_t *frame, cam_dimension_t dim,
                         }
                     }
                     CDBG_HIGH("%s: written number of bytes %ld\n",
-                             __func__, written_len);
+                             __func__, (long)written_len);
                     dumpFrmCnt++;
                     close(file_fd);
                 } else {
